+++ src/test/api/query/ant-query-manager-test.ts
@@ -5,7 +5,7 @@ import { AntPrimaryEntityManager } from '../../../persistence/primary/ant-primar
 import { SecondaryEntityManager } from '../../../persistence/secondary/secondary-entity-manager';
 import { ITest } from '../../../testapi/api/ITest';
 import { SingleResultQueryByFieldManager } from '../../primary/query/single-result-query-by-field-manager';
-import { RedisWrapper } from '../../primary/RedisWrapper';
+import { RedisWrapper } from '../../primary/redis-wrapper';
 import { MinimalAntQueryManager } from './minimal-ant-query-manager';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
