+++ src/test/primary/query/multiple-result-query-manager-test.ts
@@ -6,7 +6,7 @@ import { PrimaryEntityManager } from '../../../persistence/primary/primary-entit
 import { SecondaryEntityManager } from '../../../persistence/secondary/secondary-entity-manager';
 import { ITest } from '../../../testapi/api/ITest';
 import { SecondaryEntityManagerMock } from '../../../testapi/api/secondary/SecondaryEntityManagerMock';
-import { RedisWrapper } from '../RedisWrapper';
+import { RedisWrapper } from '../redis-wrapper';
 import { NamedEntity, NamesStartingByLetter } from './names-starting-by-letter';
 import { NamedEntityAlternative, NamesStartingByLetterAlternative } from './names-starting-by-letter-alternative';
