+++ src/test/primary/script/minimal-redis-cached-script.ts
