+++ src/test/primary/script/redis-cached-script-test.ts
@@ -1,7 +1,7 @@
 import { RedisCachedScript } from '../../../persistence/primary/script/redis-cached-script';
 import { ITest } from '../../../testapi/api/ITest';
-import { RedisWrapper } from '../RedisWrapper';
-import { MinimalRedisCachedScript } from './MinimalRedisCachedScript';
+import { RedisWrapper } from '../redis-wrapper';
+import { MinimalRedisCachedScript } from './minimal-redis-cached-script';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
