+++ src/test/primary/script/update-entities-cached-script-set-test.ts
@@ -2,7 +2,7 @@ import { AntJsUpdateOptions } from '../../../persistence/primary/options/antjs-u
 import { RedisCachedScript } from '../../../persistence/primary/script/redis-cached-script';
 import { UpdateEntitiesCachedScriptSet } from '../../../persistence/primary/script/update-entities-cached-script-set';
 import { ITest } from '../../../testapi/api/ITest';
-import { RedisWrapper } from '../RedisWrapper';
+import { RedisWrapper } from '../redis-wrapper';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
